transcribe_audio_openai_api/requirements.txt
openai
openai-whisper
python-dotenv
